MODIFY : placeTest entity to place entity
Former-commit-id: d389dd24414c7565135b45020c82dea4f07d6ba5

+++ BE/pijja/src/main/java/com/A605/pijja/domain/plan/dto/request/TmapRequestDto.java
@@ -7,8 +7,8 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 public class TmapRequestDto {
-    private float endX; //lon이 X
-    private float endY; //lat이 Y
-    private float startX;
-    private float startY;
+    private double endX; //lon이 X
+    private double endY; //lat이 Y
+    private double startX;
+    private double startY;
 }
